replay: Add --proxy command-line flag to control PROXY v2 behavior
This commit adds a --proxy command-line flag to the dnstap-replay
utility and makes it optional for the DnstapHandler to prepend a PROXY
v2 header to the DNS query message. This makes dnstap-replay usable with
DNS servers that don't implement support for the PROXY v2 header.

The DnstapHandler code that extracts the required dnstap fields and
formats them into the PROXY v2 payload has been factored out into a
standalone function and the error returns from this function are now
handled with a new `dnstap_payloads_total` metric.