Rename "mismatches" to "errors" in most cases
This commit renames "mismatches" to "errors" in most cases (especially,
the /mismatches HTTP endpoint which is now /errors, as well as some of
the metrics). The term "mismatches" was being used too loosely to mean
errors in general after additional types of errors were added to the
HTTP endpoint. For instance, timeouts are a type of error that is not a
mismatch, and it's important to log these queries through what is now
the /errors endpoint, since a timeout could be caused for instance by a
nameserver crash.

The term "mismatch" is still used to specifically mean an error that is
actually a mismatch.