replay: Convert 'dscp' argument from u8 to Option<u8>
This commit converts the 'dscp' command-line argument from u8 to
Option<u8> which more accurately reflects the underlying semantics. If
'--dscp=...' was not specified, we shouldn't try to set it, and if it
was specified, it can only be specified on platforms that support it.
This also allows setting the DSCP value to 0 on platforms that support
DSCP. The value 0 is not necessarily the default for outgoing queries
depending on how the host is configured so we should allow explicitly
setting it to 0 rather than treating 0 as a sentinel value.