DnstapHandlerError: Add deserialization function
This commit adds an ad hoc function for deserializing
DnstapHandlerError's.

Since the serialized DnstapHandlerError's that get written out by
dnstap-replay need to eventually be deserialized and formatted for
display, a function is needed to reverse the serialization performed by
the DnstapHandlerError::serialize() method.

Note that the serialization/deserialization pair of functions here
doesn't quite perform a lossless 100% roundtrip. The additional String's
attached to DnstapHandlerError::Mismatch are redundant and only used for
logging purposes, so they don't get serialized out, and thus they can't
be deserialized. But the Mismatch's String's aren't needed anyway.