New utility dnstap-dump
This commit adds a new utility "dnstap-dump" which dumps a dnstap file
for display purposes. Coincidentally, the display format is (should)
also be valid YAML.

If there is more than one dnstap payload encoded in a dnstap file, they
will be delimited using YAML's "---" document separator.

The format produced by dnstap-dump is closely based on the format that
dnstap-ldns (https://github.com/dnstap/dnstap-ldns) generates, which
itself is closely based on the dnstap protobuf schema. There are a few
minor differences, so the output is not identical. In particular:

1) dnstap-dump uses the NLnet Labs 'domain' rust crate for decoding DNS
messages rather than the NLnet Labs 'ldns' C library that dnstap-ldns
uses. There isn't a ready-made function in the 'domain' crate that
corresponds to ldns's ldns_pkt2buffer_str_fmt() function that formats an
entire DNS message object into "dig-style" output, so I implemented a
function fmt_dns_message() in dnstap-dump that drives the 'domain'
crate's API to produce something similar. It's not complete, though, in
particular it doesn't decode EDNS options. There are also some minor
whitespace differences, and some disagreements (e.g. for DNS RCODE 9
ldns uses the mnemonic "NOTAUTH" while domain uses the mnemonic
"NOAUTH").

2) Because dnstap-dump is sort of a companion utility to dnstap-replay,
in that mismatches logged by dnstap-replay will need to be analyzed,
dnstap-dump outputs both the raw (hex encoded) DNS message payload as
well as the dig-style formatted version of the DNS message payload.
These are the pairs of '..._bytes' and '..._formatted' fields. This
enables more in-depth analysis of the raw DNS message payloads in the
case where the dig-style format is identical between the two payloads.

3) dnstap-replay stuffs some serialized extra information into the
dnstap 'extra' field. This is an ad hoc "escape hatch" field in the
dnstap schema with no defined format, so there is a signature written to
this field ahead of the serialized data. dnstap-dump checks for this
signature and decodes the field if it came from dnstap-replay. If
dnstap-replay encountered a mismatch when re-sending the logged query to
the server under test, this is where the mismatched response message
will be written, and a dnstap-replay / dnstap-dump user can compare the
message: response_message_formatted field against the extra:
mismatch_formatted field to determine the difference between the
original response and the re-queried response.