publish RX counter at same time as TX counter
In order to avoid the case where we must deal with discontinuous metrics in
the downstream, publish a (zero) counter to the metrics system at the same
time as we publish an inc()rementing counter for the first time that we try to
save a message for later pickup.

This avoids (unnecessary) gymnastics in the alerting system when trying to
compare present (or absent) RX metrics with present TX metrics.