replay: Initialize metrics at startup
This commit converts dnstap-replay's metrics to use lazy_static (instead
of once_cell) and prometheus-static-metric, and pre-initializes all the
metrics at startup.

The use of prometheus-static-metric is slightly more efficient (it
"reduces the amount of branching and processing needed at runtime to
collect metrics" according to the documentation), but the main value is
being able to pre-define the metric label values that will be used in
dnstap-replay (since they're known at compile time), combined with
calling lazy_static::initialize() on each metric at process startup.

This results in all the possible metrics and their label values being
written out to the /metrics endpoint. For instance, here is the output
of the /metrics endpoint from a freshly started dnstap-replay instance
that has not processed any data yet:

    # HELP dnstap_replay_channel_error_rx_total Number of error channel receives performed.
    # TYPE dnstap_replay_channel_error_rx_total counter
    dnstap_replay_channel_error_rx_total{result="success"} 0
    # HELP dnstap_replay_channel_error_tx_total Number of error channel sends performed.
    # TYPE dnstap_replay_channel_error_tx_total counter
    dnstap_replay_channel_error_tx_total{result="error"} 0
    dnstap_replay_channel_error_tx_total{result="success"} 0
    # HELP dnstap_replay_data_bytes_total Number of Frame Streams data frame bytes processed.
    # TYPE dnstap_replay_data_bytes_total counter
    dnstap_replay_data_bytes_total 0
    # HELP dnstap_replay_data_frames_total Number of Frame Streams data frames processed.
    # TYPE dnstap_replay_data_frames_total counter
    dnstap_replay_data_frames_total 0
    # HELP dnstap_replay_dns_comparisons_total Number of DNS comparison operations performed.
    # TYPE dnstap_replay_dns_comparisons_total counter
    dnstap_replay_dns_comparisons_total{result="matched"} 0
    dnstap_replay_dns_comparisons_total{result="mismatched"} 0
    # HELP dnstap_replay_dns_queries_total Number of DNS re-query operations performed.
    # TYPE dnstap_replay_dns_queries_total counter
    dnstap_replay_dns_queries_total{result="error"} 0
    dnstap_replay_dns_queries_total{result="success"} 0
    dnstap_replay_dns_queries_total{result="timeout"} 0
    # HELP dnstap_replay_dnstap_handler_internal_errors_total Number of internal errors encountered by dnstap handler.
    # TYPE dnstap_replay_dnstap_handler_internal_errors_total counter
    dnstap_replay_dnstap_handler_internal_errors_total{result="discard_non_udp"} 0
    # HELP dnstap_replay_dnstap_payloads_total Number of dnstap payloads processed.
    # TYPE dnstap_replay_dnstap_payloads_total counter
    dnstap_replay_dnstap_payloads_total{result="error"} 0
    dnstap_replay_dnstap_payloads_total{result="success"} 0

This is apparently easier to process by downstream metric consumers
compared to the previous behavior where metrics were being lazily
initialized and were not appearing at all in the /metrics endpoint
output until they had been incremented at least once.

Note: For the dnstap_replay_dns_comparisons_total metric, the "match"
and "mismatch" label values were renamed to "matched" and "mismatched".
"match" is a Rust keyword and I was not able to find a way to escape
this value.