replay: Use Arc<AtomicBool> to communicate match status
This commit introduces an Arc<AtomicBool> flag shared between the
MonitorHandler and DnstapHandler(s) that controls whether to suppress
the response matching process (if the flag is set to false) or whether
to proceed with the normal response matching process (if the flag is set
to true).

This allows coordinating external processes to provide status files that
can temporarily suppress dnstap-replay's response matching process. For
instance, if the source or target nameserver in the dnstap-replay
pipeline is reloading its configuration or data it might be useful to
temporarily suppress reloads during the reload.

This commit also introduces a `dnstap_replay_dns_comparisons_total`
metric that counts the number of suppressed comparisons.