replay: Add a "dnstap_replay_match_status" metric
This commit adds a "dnstap_replay_match_status" metric that reflects
whether match comparisons are enabled (1) or suppressed (0). This metric
will always be set to 1, unless the match status monitor is enabled on
the command-line. In that case the metric will be set to 1 or 0 as the
match status monitor detects changes to the match status files.