replay: Add a dedicated "timeouts" channel
This commit adds a dedicated "timeouts" channel used internally between
`DnstapHandler` and `HttpHandler`, which is similar to the existing
"errors" channel, and updates the `DnstapHandler` so that timeout
messages are sent via the "timeouts" channel rather than the "errors"
channel.

With this change, mismatches and timeouts will no longer be mixed
together into the same "errors" channel.

This commit also introduces a new /timeouts HTTP endpoint for retrieving
the current contents of the "timeouts" channel.

New metrics (`dnstap_replay_channel_timeout_rx_total` and
`dnstap_replay_channel_timeout_tx_total`) are introduced for monitoring
the utilization of the "timeouts" channel.