replay: Add --ignore-tc option to ignore UDP mismatches when TC bit is set
This commit adds a `--ignore-tc` command-line option that controls the
behavior of the DnstapHandler's response message matching algorithm.

When enabled, this option avoids a type of nuisance mismatch where RRL
is triggered and causes the TC bit to be set on a response. Since RRL is
timing dependent and dnstap is heavily buffered, dnstap-replay may not
replicate the exact same pattern of RRL-controlled responses, resulting
in a mismatch.

If `--ignore-tc` is enabled and the pair of DNS response messages being
analyzed don't match exactly, the response message headers will be
further tested to see if one of the message headers has the TC bit set.
If so, instead of being treated as a mismatch, a separate
`udp_tc_ignored` metric will be incremented instead.