replay: Fix confusion between error/timeout channel metrics
When the dedicated /timeouts endpoint was introduced in
https://github.com/fastly/dnstap-utils/pull/22 to split out the timeout
payloads from the mismatch payloads on the existing /errors endpoint,
new metrics were introduced for monitoring the utilization of the
corresponding internal "timeouts" channel.

However, the internal `dnstap_receiver_to_stream()` function in the HTTP
handler wasn't updated due to this change and it continued to increment
the metric for RX's from the "errors" channel regardless of whether
payloads were being received from the "errors" or the newly introduced
"timeouts" channel.

This commit fixes the metric increment in the HTTP handler by
encapsulating the channel to receive from along with the metric to
increment on a successful RX together into a structure and passing this
structure down to the `dnstap_receiver_to_stream()` function.