fmt_dns_message(): Try to print most of the EDNS options
This commit enhances `fmt_dns_message()` to print most of the known EDNS
options in an ad hoc format.

This is a little bit fragile. If the EDNS option is malformed in some
way, the opt record simply won't be present in the parsed message
object, and nothing will be printed for the OPT pseudosection at all.
But this is similar to the case today where the options are entirely
ignored.

We might be able to simplify this code in an upcoming version of the
`domain` crate (https://github.com/NLnetLabs/domain/issues/144).