replay/dnstap_handler: Factor out `query_address` extraction
This commit factors out the extraction and conversion of the dnstap
`query_address` field from the `add_proxy_payload()` helper function
into its caller `DnstapHandler::process_dnstap_message()`. The result of
this extraction will need to be reused when query net filtering is
added.

Note that this changes the behavior of
`DnstapHandler::process_dnstap_message()` slightly. Previously it was
not mandatory for a dnstap Message payload to have a `query_address`
field *unless* dnstap-replay is running with the `--proxy` command-line
option enabled. This change now makes it mandatory for the dnstap
message payload to have the `query_address` field present regardless of
whether `--proxy` is set or not. This is not that significant of a
change since dnstap Message payloads should generally have the
`query_address` field set anyway, especially the `*_RESPONSE` Types.