replay: Add `--proxy-timespec` command-line flag to add custom PROXY v2 timespec TLV
This commit adds a `--proxy-timespec` command-line flag to the
dnstap-replay utility that controls the behavior of the PROXY v2
encoder (if `--proxy` is also set).

The PROXY v2 protocol supports [encoding additional TLVs] after the
address block:

    If the length specified in the PROXY protocol header indicates that
    additional bytes are part of the header beyond the address
    information, a receiver may choose to skip over and ignore those
    bytes, or attempt to interpret those bytes.

    The information in those bytes will be arranged in Type-Length-Value
    (TLV vectors) in the following format. The first byte is the Type of
    the vector. The second two bytes represent the length in bytes of
    the value (not included the Type and Length bytes), and following
    the length field is the number of bytes specified by the length.

            struct pp2_tlv {
                uint8_t type;
                uint8_t length_hi;
                uint8_t length_lo;
                uint8_t value[0];
            };

    A receiver may choose to skip over and ignore the TLVs it is not
    interested in or it does not understand. Senders can generate the
    TLVs only for the information they choose to publish.

Additional, there is a range of TLV codepoints [reserved for
application-specific data]:

    The following range of 16 type values is reserved for
    application-specific data and will be never used by the PROXY
    Protocol. If you need more values consider extending the range with
    a type field in your TLVs.

            #define PP2_TYPE_MIN_CUSTOM    0xE0
            #define PP2_TYPE_MAX_CUSTOM    0xEF

This commit uses the codepoint 0xEA for dnstap-replay's PROXY v2
timespec TLV. This TLV represents the timestamp at which the original
DNS response was processed. This allows the DNS server under test
(receiving replayed DNS queries with prepended PROXY v2 headers) to
recover the original timestamp of the DNS query, which allows for
testing time-dependent behavior. Because of buffering and processing
latency, etc. there can be a small but non-negligible amount of delay
(up to a few seconds) between when the original DNS server responds to
the original query and when the DNS server under test processes the
replayed DNS query.

The timespec TLV encodes a "timespec"-compatible structure which
represents the number of seconds elapsed since the Unix epoch,
represented as a combination of a whole number (seconds) and fractional
part (nanoseconds). The whole number is represented as a little endian
64-bit integer and the fractional part is represented as a little endian
32-bit integer.

[encoding additional TLVs]: https://github.com/haproxy/haproxy/blob/v2.7.0/doc/proxy-protocol.txt#L514-L534
[reserved for application-specific data]: https://github.com/haproxy/haproxy/blob/v2.7.0/doc/proxy-protocol.txt#L677-L682