Replace `chrono` with `time` for timestamp formatting
This commit uses the `time` crate to perform timestamp formatting
instead of `chrono`. There does not seem to be an exact match in the
`time` crate for formatting timestamps the way we were doing it before
(basically "%Y-%m-%d %H:%M:%S" in strftime syntax) so I wrote a helper
function that uses a custom time format description.

This is the one place where we directly use the `chrono` crate.